Fix for max-count option
git-svn-id: http://parabix.costar.sfu.ca/svn/icGREP/icgrep-devel@5477 a82c6ad6-7e1c-4be0-8e68-26c050b43eaf

+++ icgrep/grep_interface.h
@@ -75,10 +75,10 @@ extern bool InitialTabFlag; // -T
 extern bool OnlyMatchingFlag; // -o
 extern bool LineBufferedFlag; // -line-buffered
 extern bool NormalizeLineBreaksFlag; // -normalize-line-breaks
-extern size_t AfterContextFlag; // -A
-extern size_t BeforeContextFlag; // -B
-extern size_t ContextFlag; // -C
-extern size_t MaxCountFlag; // -m
+extern int AfterContextFlag; // -A
+extern int BeforeContextFlag; // -B
+extern int ContextFlag; // -C
+extern int MaxCountFlag; // -m
     
 
 //
